Main Script
First, the data is imported.
Categorical features were eliminated from both training and test data, however, they can be included by using one hot encoding. In this implementation, that was not done.
The Not-a-Number values were filled with the median of that feature.
I decided to use Random Forest Estimator because the features of the data were diverse.
Then a Random Forest Classifier is trained using grid search with cross validation.
Number of Trees and Maximum Depth parameters of the Random Forest was detected using gridsearchCV() function, respectively, in the range 70-120 and 5-10, for all 3 data sets.